Need sleep...
Still missing:
  - Converting all value-fetchers to correct method-usage, to avoid stupid null values.
  - Test that simple connections are functioning.
  - Finish the simple connections once confirmed.
  - Test complex connections.
  - Implement these fully.

+++ McSntt/McSntt/DataAbstractionLayer/BoatEfDal.cs
@@ -73,7 +73,7 @@ namespace McSntt.DataAbstractionLayer
             }
         }
 
-        public Boat GetOne(int itemId)
+        public Boat GetOne(long itemId)
         {
             using (var db = new McSntttContext())
             {
