+++ McSntt/McSntt/DataAbstractionLayer/IGenericDal.cs
@@ -12,6 +12,6 @@ namespace McSntt.DataAbstractionLayer
         bool Update(params T[] items);
         bool Delete(params T[] items);
         IEnumerable<T> GetAll();
-        T GetOne(int itemId);
+        T GetOne(long itemId);
     }
 }
